pose_backend/requirements.txt
fastapi
uvicorn
opencv-python
mediapipe
numpy
pillow
